refactor v1 trace lazy operators
